requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
python-dotenv==1.0.1
pytest==8.0.2
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
requests==2.31.0
psutil==5.9.8  # For system metrics
SQLAlchemy==1.4.49  # For database operations
flask_sqlalchemy==2.5.1  # For Flask-SQLAlchemy integration
marshmallow==3.21.0  # For schema validation
Jinja2==3.1.3  # For HTML templates 